Some more effort at making Bun modular
