+++ Bun/Bun.vcxproj
@@ -168,6 +168,10 @@
   <ItemGroup>
     <ClInclude Include="..\include\blib\bun\Bun.hpp" />
     <ClInclude Include="..\include\blib\bun\CppTypeToSQLString.hpp" />
+    <ClInclude Include="..\include\blib\bun\DbBackend.hpp" />
+    <ClInclude Include="..\include\blib\bun\DbLogger.hpp" />
+    <ClInclude Include="..\include\blib\bun\PRef.hpp" />
+    <ClInclude Include="..\include\blib\bun\Query.hpp" />
     <ClInclude Include="..\include\blib\bun\SimpleOID.hpp" />
     <ClInclude Include="..\include\blib\containers\tree\NTree.hpp" />
     <ClInclude Include="..\include\blib\fmt\format.hpp" />
