+++ include/blib/bun/Bun.hpp
@@ -11,16 +11,15 @@
 #include "blib/utils/MD5.hpp"
 #include "blib/utils/TypeUtils.hpp"
 #include "blib/bun/CppTypeToSQLString.hpp"
-#include "blib/bun/SimpleOID.hpp"
-#include <blib/hdr/sqlite_modern_cpp.hpp>
+#include "blib/bun/Query.hpp"
+#include "blib/bun/PRef.hpp"
+#include "blib/bun/DbBackend.hpp"
+#include "blib/bun/DbLogger.hpp"
 #include <boost/preprocessor.hpp>
-#include <boost/proto/proto.hpp>
 #include <blib/fmt/format.hpp>
-#include <blib/spdlog/spdlog.h>
 #include <blib/rapidjson/rapidjson.h>
 #include <memory>
 #include <tuple>
-#include <bitset>
 #include <array>
 #include <ratio>
 #include <stack>
@@ -29,57 +28,6 @@
 
 namespace blib {
   namespace bun {
-    /////////////////////////////////////////////////
-    /// @brief Returns the logger
-    ///        This will be logging to the logs. It uses spdlog to log to file
-    /// @return Returns the logger.
-    /////////////////////////////////////////////////
-    inline spdlog::logger& l() {
-      static const std::size_t q_size = 1048576; //queue size must be power of 2
-      spdlog::set_async_mode( q_size );
-      static auto ret = spdlog::daily_logger_st( "async_file_logger",
-        "query_log.txt" );
-      return *ret;
-    }
-
-    /////////////////////////////////////////////////
-    /// @class Db
-    /// @brief Database class. This is a singleton class.
-    ///        This class is the backend that does the persist job.
-    ///        This class should not be used directly.
-    /////////////////////////////////////////////////
-    struct Db : public ::blib::Singleton<Db> {
-      using DbConnectionType = sqlite::database;
-      std::unique_ptr<DbConnectionType> _db;
-      bool _ok;
-      std::string _fileName;
-
-      bool ok() const {
-        return _ok;
-      }
-
-      DbConnectionType& db() {
-        return *_db;
-      }
-
-      bool connect( std::string const& aDbName ) {
-        bool ret = true;
-        try {
-          _db.reset( new DbConnectionType( aDbName ) );
-          _ok = true;
-          _fileName = aDbName;
-        }
-        catch (...) {
-          ret = false;
-          _ok = false;
-        }
-        return ret;
-      }
-    };
-
-    inline void dbConnect( std::string const& in_db ) {
-      Db::i().connect( in_db );
-    }
 
     /////////////////////////////////////////////////
     /// @class BunHelper
@@ -107,205 +55,10 @@ namespace blib {
     template<typename T>
     inline std::vector<SimpleOID> getAllOids();
 
-    /////////////////////////////////////////////////
-    /// @class PRef
-    /// @brief The persistent reference holder.
-    ///        This is the primary object holder.
-    /////////////////////////////////////////////////
-    template<typename T>
-    class PRef {
-    private:
-      enum class FlagsE : std::uint8_t {
-        kDirty = 0
-      };
-
-      typedef SimpleOID OidType;
-      std::unique_ptr<T> _obj;
-      std::bitset<4> _flags;
-      std::string _md5;
-
-    public:
-      typedef T ObjType;
-      typedef PRef<T> SelfType;
-      OidType oid;
-
-    public:
-      PRef() = default;
-      PRef( PRef const& in_other ) : oid( in_other.oid ) {
-        load( oid );
-      }
-
-      PRef( PRef& in_other ) : oid( in_other.oid ), _flags( in_other._flags ), _md5( in_other._md5 ), _obj( in_other._obj.release() ) {
-      }
-
-      PRef( ObjType* in_obj ) :_obj( in_obj ) {
-      }
-
-      PRef( OidType const& in_oid ) : oid( in_oid ) {
-        load( oid );
-      }
-
-      PRef( OidType const& in_oid, ObjType* in_obj ) : oid( in_oid ), _obj( in_obj ) {
-      }
-
-      void reset( ObjType* in_obj ) {
-        _obj.reset( in_obj );
-        _flags.reset();
-        _md5 = "";
-      }
-
-      ObjType* release() {
-        oid.high = 0;
-        oid.low = 0;
-        _flags.reset();
-        _md5 = "";
-        return _obj.release();
-      }
-
-      ~PRef() = default;
-
-      decltype(*_obj) operator*() {
-        return *_obj;
-      }
-
-      T* operator->() {
-        return _obj.get();
-      }
-
-      bool dirty() {
-        const auto md5 = BunHelper<ObjType>::md5( _obj.get(), oid );
-        if (md5 != _md5) {
-          _flags[static_cast<std::uint8_t>(FlagsE::kDirty)] = 1;
-        }
-
-        return _flags[static_cast<std::uint8_t>(FlagsE::kDirty)] ? true : false;
-      }
-
-      OidType persist() {
-        if (_md5.empty()) {
-          oid = BunHelper<ObjType>::persistObj( _obj.get() );
-        }
-        else {
-          BunHelper<ObjType>::updateObj( _obj.get(), oid );
-        }
-        _md5 = BunHelper<ObjType>::md5( _obj.get(), oid );
-        _flags.reset();
-        return oid;
-      }
-
-      OidType save() {
-        return persist();
-      }
-
-      void del() {
-        BunHelper<ObjType>::deleteObj( oid );
-        _md5.clear();
-        oid.clear();
-        _flags.reset();
-      }
-
-      PRef& operator=( ObjType* in_obj ) {
-        reset( in_obj );
-        return *this;
-      }
-
-      PRef& operator=( PRef& in_other ) {
-        copyFrom( in_other );
-        return *this;
-      }
-
-      bool operator==( PRef const& in_other ) {
-        return oid == in_other.oid;
-      }
-
-      bool operator!=( PRef const& in_other ) {
-        return oid != in_other.oid;
-      }
-
-      std::string toJson() const {
-        return BunHelper<T>::objToJson( _obj.get(), oid );
-      }
-
-    private:
-      void load( OidType const& in_oid ) {
-        oid = in_oid;
-        _obj = BunHelper<ObjType>::getObj( oid );
-        _md5 = BunHelper<ObjType>::md5( _obj.get(), oid );
-        _flags.reset();
-      }
-
-      void copyFrom( PRef& in_other ) {
-        oid = in_other.oid;
-        _md5 = in_other._md5;
-        _flags = in_other._flags;
-        _obj = in_other._obj;
-      }
-    };
-
-    namespace _details {
-      template<typename T>
-      struct BoxingType;
-
-      template<>
-      struct BoxingType<int> {
-        using type = std::int64_t;
-      };
-
-      template<>
-      struct BoxingType<unsigned int> {
-        using type = std::uint64_t;
-      };
-
-      template<>
-      struct BoxingType<float> {
-        using type = double;
-      };
-
-      template<>
-      struct BoxingType<double> {
-        using type = double;
-      };
-
-      template<>
-      struct BoxingType<char> {
-        using type = std::int64_t;
-      };
-
-      template<>
-      struct BoxingType<unsigned char> {
-        using type = std::uint64_t;
-      };
-
-      template<>
-      struct BoxingType<std::string> {
-        using type = std::string;
-      };
-
-      /////////////////////////////////////////////////
-      /// @fn _E
-      /// @brief Method to put quote around strings. This is used for the 
-      ///        formation of the Sql String.
-      /////////////////////////////////////////////////
-      template<typename T>
-      inline T _E( const T in_val ) {
-        return in_val;
-      }
-
-      inline std::string& _E( std::string& in_val ) {
-        in_val = "'" + in_val + "'";
-        return in_val;
-      }
-    }
-
     template<typename T>
     inline std::vector<PRef<T>> getAllObjects();
-  }
-}
 
-///------------------------------------------------------------------------
-/// Bun Query Start
-namespace blib {
-  namespace bun {
+    /// Query namespace
     namespace query {
       namespace _details {
         template<typename T>
@@ -313,283 +66,11 @@ namespace blib {
 
         template<typename T>
         inline std::string const& className();
-
-        template<std::int32_t I>
-        struct QueryVariablePlaceholderIndex : std::integral_constant <std::int32_t, I> {
-        };
-
-        // Grammar for the query Start
-        struct PlaceHoldersTerminals : boost::proto::or_ <
-          boost::proto::terminal<QueryVariablePlaceholderIndex<0>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<1>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<2>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<3>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<4>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<5>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<6>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<7>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<8>>,
-          boost::proto::or_<
-          boost::proto::terminal<QueryVariablePlaceholderIndex<9>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<10>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<11>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<12>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<13>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<14>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<15>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<16>>,
-          boost::proto::terminal<QueryVariablePlaceholderIndex<17>>
-          >
-        > {
-        };
-
-        struct QueryLiteralTerminals : boost::proto::or_<
-          boost::proto::terminal<boost::proto::convertible_to<int>>,
-          boost::proto::terminal<boost::proto::convertible_to<unsigned int>>,
-          boost::proto::terminal<boost::proto::convertible_to<char>>,
-          boost::proto::terminal<boost::proto::convertible_to<unsigned char>>,
-          boost::proto::terminal<boost::proto::convertible_to<float>>,
-          boost::proto::terminal<boost::proto::convertible_to<double>>,
-          boost::proto::terminal<boost::proto::convertible_to<bool>>,
-          boost::proto::terminal<boost::proto::convertible_to<std::string>>
-        > {
-        };
-
-        struct AllTerminals : boost::proto::or_<
-          PlaceHoldersTerminals,
-          QueryLiteralTerminals
-        > {
-        };
-
-        struct GreaterSymbols : boost::proto::or_<
-          boost::proto::greater<AllTerminals, AllTerminals>,
-          boost::proto::greater_equal<AllTerminals, AllTerminals>
-        > {
-        };
-
-        struct LessSymbols : boost::proto::or_<
-          boost::proto::less<AllTerminals, AllTerminals>,
-          boost::proto::less_equal<AllTerminals, AllTerminals>
-        > {
-        };
-
-        struct EqualSymbols : boost::proto::or_<
-          boost::proto::equal_to<AllTerminals, AllTerminals>,
-          boost::proto::not_equal_to<AllTerminals, AllTerminals>
-        > {
-        };
-
-        struct BunQueryGrammar : boost::proto::or_<
-          AllTerminals,
-          GreaterSymbols,
-          LessSymbols,
-          EqualSymbols,
-          boost::proto::logical_and<BunQueryGrammar, BunQueryGrammar>,
-          boost::proto::logical_or<BunQueryGrammar, BunQueryGrammar>
-        > {
-        };
-
-        // Grammar for the query End
-        template<typename T>
-        inline std::string const& mapping( const std::uint32_t in_index );
-        //{
-        //  static const std::vector<std::string> ret = { "name", "age", "height" };
-        //  return ret.at( in_index );
-        //}
-
-        template<typename T>
-        struct TypesUsed;
-        //boost::mpl::vector<decltype(test::Person::name), decltype(test::Person::age), decltype(test::Person::height)>;
-
-        template<typename T>
-        struct FromInternals {
-          using TypesUsed = typename TypesUsed<T>::Type;
-
-          struct BunQueryFilterContex : boost::proto::callable_context<BunQueryFilterContex> {
-            typedef std::string result_type;
-            BunQueryFilterContex() {}
-
-            template<typename T>
-            result_type operator ()( boost::proto::tag::terminal, T in_term ) const {
-              const auto ret = std::to_string( blib::bun::_details::_E( in_term ) );
-              return ret;
-            }
-
-            result_type operator ()( boost::proto::tag::terminal, std::string& in_term ) const {
-              const auto ret = blib::bun::_details::_E( in_term );
-              return ret;
-            }
-
-            result_type operator ()( boost::proto::tag::terminal, char const* in_term ) const {
-              std::string str = std::string( in_term );
-              const auto ret = blib::bun::_details::_E( str );
-              return ret;
-            }
-
-            template<std::uint32_t I>
-            result_type operator()( boost::proto::tag::terminal, bun::query::_details::QueryVariablePlaceholderIndex<I> in_term ) const {
-              const auto ret = mapping<T>( I );
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::logical_and, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " AND ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::logical_or, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " OR ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::less, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " < ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::less_equal, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " <= ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::greater, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " > ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::greater_equal, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " >= ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::equal_to, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " == ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename L, typename R>
-            result_type operator()( boost::proto::tag::not_equal_to, L const& in_l, R const& in_r ) const {
-              static const std::string operator_name = " != ";
-              auto ctx = *this;
-              const auto left_string = boost::proto::eval( in_l, ctx );
-              const auto right_string = boost::proto::eval( in_r, ctx );
-              const std::string ret = left_string + operator_name + right_string;
-              return ret;
-            }
-
-            template<typename TerminalType, typename L, typename R>
-            result_type operator()( TerminalType, L const& in_l, R const& in_r ) const {
-              static_assert(false, "Operator not supported in Bun");
-              const std::string ret = "Operator not supported in Bun";
-              return ret;
-            }
-
-            template<typename TerminalType, typename L>
-            result_type operator()( TerminalType, L const& in_l ) const {
-              static_assert(false, "Operator not supported in Bun");
-              const std::string ret = "Operator not supported in Bun";
-              return ret;
-            }
-          };
-        };
       }
-
-      // Fields for query
-      namespace {
-        template<typename T>
-        struct F;
-      }
-
-      /////////////////////////////////////////////////
-      /// ... text ...
-      /////////////////////////////////////////////////
-      template<typename T>
-      struct From {
-      private:
-        std::string _query;
-        decltype(_details::getAllObjWithQuery<T>( "" )) _objects;
-
-      private:
-        template<typename ExpressionType>
-        std::string eval( ExpressionType const& in_expr ) {
-          _details::FromInternals<T>::BunQueryFilterContex ctx;
-          const std::string ret = boost::proto::eval( in_expr, ctx );
-          return ret;
-        }
-
-        static std::string const& className() {
-          static const auto table_name = className<T>();
-          return table_name;
-        }
-
-      public:
-        From() = default;
-
-        From( From& in_other ) :_query( in_other._query ), _objects( in_other._objects ) {}
-
-        template<typename ExpressionType>
-        From& where( ExpressionType const& in_expr ) {
-          static_assert(boost::proto::matches<ExpressionType, _details::BunQueryGrammar>::value, "Syntax error in Bun Query");
-          const std::string query_string = eval( in_expr );
-          const std::string add_string = _query.empty() ? "" : " AND ";
-          _query += add_string + query_string;
-          return *this;
-        }
-
-        std::string const& query() const {
-          l().info() << _query;
-          return _query;
-        }
-
-        decltype(_objects)& objects() {
-          _objects = _details::getAllObjWithQuery<T>( _query );
-          return _objects;
-        }
-      };
-
-      template<typename ExpressionType>
-      struct IsValidQuery : std::integral_constant<bool, boost::proto::matches<ExpressionType, _details::BunQueryGrammar>::value> {
-      };
     }
   }
 }
-/// Bun Query End
-///------------------------------------------------------------------------
+
 
 ///----------------------------------------------------------------------------
 /// Basic Persistance Start
